+++ application-domain/src/main/kotlin/hs/kr/equus/application/domain/application/spi/PrintApplicantCodesPort.kt
@@ -1,6 +1,6 @@
 package hs.kr.equus.application.domain.application.spi
 
-import hs.kr.equus.application.domain.application.usecase.dto.response.ApplicationCodeVO
+import hs.kr.equus.application.domain.application.usecase.dto.vo.ApplicationCodeVO
 import javax.servlet.http.HttpServletResponse
 
 interface PrintApplicantCodesPort {
