+++ application-domain/src/main/kotlin/hs/kr/equus/application/domain/application/usecase/dto/vo/ApplicationCodeVO.kt
@@ -1,4 +1,4 @@
-package hs.kr.equus.application.domain.application.usecase.dto.response
+package hs.kr.equus.application.domain.application.usecase.dto.vo
 
 data class ApplicationCodeVO (
     val receiptCode: Long,
