+++ application-infrastructure/src/main/kotlin/hs/kr/equus/application/domain/application/domain/ApplicationPersistenceAdapter.kt
@@ -9,7 +9,7 @@ import hs.kr.equus.application.domain.application.model.types.ApplicationType
 import hs.kr.equus.application.domain.application.spi.ApplicationPort
 import hs.kr.equus.application.domain.application.usecase.dto.response.GetApplicationCountResponse
 import hs.kr.equus.application.domain.application.usecase.dto.response.GetStaticsCountResponse
-import hs.kr.equus.application.domain.application.usecase.dto.response.ApplicationCodeVO
+import hs.kr.equus.application.domain.application.usecase.dto.vo.ApplicationCodeVO
 import hs.kr.equus.application.global.feign.client.StatusClient
 import hs.kr.equus.application.global.feign.client.dto.response.StatusInfoElement
 import org.springframework.stereotype.Component
