+++ application-infrastructure/src/main/kotlin/hs/kr/equus/application/global/excel/generator/PrintApplicantCodesGenerator.kt
@@ -1,7 +1,7 @@
 package hs.kr.equus.application.global.excel.generator
 
 import hs.kr.equus.application.domain.application.spi.PrintApplicantCodesPort
-import hs.kr.equus.application.domain.application.usecase.dto.response.ApplicationCodeVO
+import hs.kr.equus.application.domain.application.usecase.dto.vo.ApplicationCodeVO
 import hs.kr.equus.application.global.excel.ApplicantCode
 import hs.kr.equus.application.global.excel.exception.ExcelExceptions
 import org.apache.poi.ss.usermodel.Row
